+++ CHANGES.md
@@ -1,5 +1,5 @@
 #### [unreleased]
-* use `is_localhost()` for local git server setting
+* use `is_localhost()` for local git server setting and make automatic
 
 #### 2.5.1 / 2020-01-24
 * run if no settings yet saved